requirements.txt
python-telegram-bot[job-queue]
pandas
openpyxl
gspread
google-auth
python-dotenv